+++ packages/fal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/fal
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [f10f0fa]
+  - @ai-sdk/provider-utils@2.2.1
+
 ## 0.1.0
 
 ### Minor Changes